Add the handleRemoveBook function and pass it as a prop
